pyproject.toml
[project]
name = "Haruki-Sekai-Asset-Updater"
version = "3.0.0"
description = "A game asset extractor with HTTP API for Project Sekai"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "acb-py>=1.2.4",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.12",
    "coloredlogs>=15.0.1",
    "cryptography>=45.0.4",
    "hypercorn>=0.17.3",
    "msgpack>=1.1.0",
    "pycricodecs",
    "orjson>=3.10.18",
    "pydantic>=2.10.6",
    "pydub>=0.25.1",
    "quart>=0.20.0",
    "rich~=14.0.0",
    "ruff>=0.11.13",
    "tenacity~=9.1.2",
    "unitypy~=1.22.5",
]

[tool.uv.sources]
pycricodecs = { git = "https://github.com/Youjose/PyCriCodecs" }

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "W", # pycodestyle warnings
    "E", # pycodestyle errors
    "I", # isort
    "UP", # pyupgrade
    "ASYNC", # flake8-async
    "C4", # flake8-comprehensions
    "T10", # flake8-debugger
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "Q", # flake8-quotes
    "TID", # flake8-tidy-imports
    "RUF", # Ruff-specific rules
]
ignore = [
    "E402", # module-import-not-at-top-of-file
    "UP037", # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["nonebot", "tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true